// frontend/src/shared/api/generated/schemas/projectSubscriptionAccessResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Bab API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectSubscriptionAccessResponse {
  id: string;
  org_id: string;
  project_id: string;
  subscription_id: string;
  priority: number;
  is_active: boolean;
  created_at: string;
  updated_at: string;
}